Create globals module
